requirements.txt
requests==2.32.5
pandas==2.3.3
python-dotenv==1.1.1